+++ client/app/styles/_contrast-overrides.scss
@@ -18,6 +18,7 @@ $node-highlight-stroke-width: 0.16;
 $node-border-stroke-width: 0.2;
 $node-shadow-stroke-width: 0.25;
 $node-pseudo-opacity: 1;
+$node-blur-opacity: 0.6;
 $edge-highlight-opacity: 0.3;
 $edge-opacity-blurred: 0;
