+++ client/app/styles/_variables.scss
@@ -36,6 +36,7 @@ $node-highlight-stroke-width: 0.04;
 $node-border-stroke-width: 0.12;
 $node-shadow-stroke-width: 0.18;
 $node-pseudo-opacity: 0.8;
+$node-blur-opacity: 0.25;
 $node-text-scale: 2;
 $edge-highlight-opacity: 0.1;
 $edge-opacity-blurred: 0.2;
